Enable embeddings for append streams
