+++ crates/data_components/src/debezium/arrow/changes.rs
@@ -14,9 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use std::sync::Arc;
-
-use crate::cdc::ChangeBatchError;
 use crate::{
     arrow::struct_builder::StructBuilder,
     cdc::{ChangeBatch, changes_schema},
@@ -26,42 +23,10 @@ use crate::{
     },
 };
 use arrow::{
-    array::{Array, ArrayBuilder, ListBuilder, RecordBatch, StringBuilder, StructArray},
+    array::{ArrayBuilder, ListBuilder, RecordBatch, StringBuilder},
     datatypes::SchemaRef,
 };
 use snafu::prelude::*;
-
-pub fn replace_change_batch_data(
-    new_data: &RecordBatch,
-    change: &ChangeBatch,
-) -> Result<ChangeBatch, ChangeBatchError> {
-    let schema = changes_schema(&new_data.schema());
-
-    let cols = change
-        .record
-        .schema()
-        .fields()
-        .iter()
-        .map(|f| {
-            if f.name() == "data" {
-                Arc::new(StructArray::new(
-                    new_data.schema().fields().clone(),
-                    new_data.columns().to_vec(),
-                    None,
-                )) as Arc<dyn Array>
-            } else {
-                match change.record.column_by_name(f.name()) {
-                    Some(column) => Arc::clone(column),
-                    None => unreachable!("Column {} must exist", f.name()),
-                }
-            }
-        })
-        .collect();
-
-    RecordBatch::try_new(schema.into(), cols)
-        .map_err(|source| ChangeBatchError::Arrow { source })
-        .and_then(ChangeBatch::try_new)
-}
 
 /// Converts a `ChangeEvent` into a `ChangeBatch`
 pub fn to_change_batch(
